+++ components/Form/Form.jsx
@@ -2,8 +2,10 @@ import styled from '@emotion/styled'
 
 const Form = styled.form`
   .chakra-form-control {
-    display: grid;
-    grid-template-columns: [label] 150px [control] auto;
+    @media (min-width: 768px) {
+      display: grid;
+      grid-template-columns: [label] 150px [control] auto;
+    }
     margin-bottom: 1em;
   }
 
@@ -12,5 +14,9 @@ const Form = styled.form`
     padding-top: 0.5em;
   }
 `
+
+Form.defaultProps = {
+  method: 'post',
+}
 
 export default Form