+++ components/Map/Map.jsx
@@ -9,10 +9,28 @@ import {
 import PropTypes from 'prop-types'
 import { useCallback, useState } from 'react'
 
+function defaultIcon(props = {}) {
+  return {
+    path:
+      'M25 0c-8.284 0-15 6.656-15 14.866 0 8.211 15 35.135 15 35.135s15-26.924 15-35.135c0-8.21-6.716-14.866-15-14.866zm-.049 19.312c-2.557 0-4.629-2.055-4.629-4.588 0-2.535 2.072-4.589 4.629-4.589 2.559 0 4.631 2.054 4.631 4.589 0 2.533-2.072 4.588-4.631 4.588z',
+    fillColor: 'green',
+    fillOpacity: 1,
+    strokeColor: '#000',
+    strokeWeight: 1,
+    scale: 0.5,
+    anchor: new window.google.maps.Point(10, 30),
+    ...props,
+  }
+}
+function unavailableIcon() {
+  return defaultIcon({
+    fillColor: 'gray',
+  })
+}
+
 const Map = ({ markers, zoom, onRenderWindow, ...props }) => {
   const [selectedPlace, setSelectedPlace] = useState(null)
   const [markerMap, setMarkerMap] = useState({})
-  const [center, setCenter] = useState({})
   const [infoOpen, setInfoOpen] = useState(false)
 
   const { isLoaded } = useLoadScript({
@@ -34,7 +52,6 @@ const Map = ({ markers, zoom, onRenderWindow, ...props }) => {
 
   const markerClickHandler = (place) => {
     setSelectedPlace(place)
-    setCenter(place.pos)
 
     if (!onRenderWindow) return
 
@@ -55,7 +72,6 @@ const Map = ({ markers, zoom, onRenderWindow, ...props }) => {
     return (
       <GoogleMap
         onLoad={onLoad}
-        center={center}
         zoom={zoom}
         mapContainerStyle={{
           height: '100%',
@@ -66,8 +82,10 @@ const Map = ({ markers, zoom, onRenderWindow, ...props }) => {
           <Marker
             key={place.id}
             position={place.pos}
+            icon={!place.selected ? defaultIcon() : unavailableIcon()}
             onLoad={(marker) => markerLoadHandler(marker, place)}
-            onClick={() => markerClickHandler(place)}
+            clickable={!place.selected}
+            onClick={() => !place.selected && markerClickHandler(place)}
           />
         ))}
         {infoOpen && selectedPlace && (
@@ -84,7 +102,7 @@ const Map = ({ markers, zoom, onRenderWindow, ...props }) => {
 
   return (
     <Skeleton isLoaded={isLoaded}>
-      <Box height="100vh" width="100%" {...props}>
+      <Box height={['100vh', '75vh']} width="100%" {...props}>
         {isLoaded ? renderMap() : null}
       </Box>
     </Skeleton>