+++ components/Modal/Modal.jsx
@@ -8,7 +8,7 @@ import {
 import PropTypes from 'prop-types'
 
 const Modal = ({ children, ...props }) => (
-  <ChakraModal {...props}>
+  <ChakraModal {...props} size={['full', '3xl']}>
     <ModalOverlay />
     <ModalContent>
       <ModalCloseButton />