+++ pages/_app.js
@@ -1,7 +1,21 @@
-import '../styles/globals.css'
+import { configureStore } from '@reduxjs/toolkit'
+import Layout from 'components/Layout'
+import { Provider } from 'react-redux'
+import rootReducer from 'reducers'
 
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+const store = configureStore({
+  reducer: rootReducer,
+})
+
+// eslint-disable-next-line
+function App({ Component, pageProps }) {
+  return (
+    <Provider store={store}>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </Provider>
+  )
 }
 
-export default MyApp
+export default App