+++ pages/success.jsx
@@ -1,4 +1,4 @@
-import { Flex, Link } from '@chakra-ui/react'
+import { Flex, Link, Stack } from '@chakra-ui/react'
 import Head from 'next/head'
 import NextLink from 'next/link'
 
@@ -9,9 +9,14 @@ export default function Home() {
         <title>Distribute Product Order</title>
       </Head>
       <Flex align="center" justifyContent="center" height="100vh">
-        <NextLink href="/cart">
-          <Link>Go To Cart</Link>
-        </NextLink>
+        <Stack>
+          <NextLink href="/cart">
+            <Link>Go To Cart</Link>
+          </NextLink>
+          <NextLink href="/">
+            <Link>Go To Home</Link>
+          </NextLink>
+        </Stack>
       </Flex>
     </>
   )