+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.4](///compare/v0.1.3...v0.1.4) (2021-03-20)
+
+
+### Features
+
+* add `/cart` e342ca4
+* add `/success` 8eed20b
+* add `cartSlice` 252e000
+* add `DateSelector` ff46398
+* add `locationSlice` 21c668d
+* add `ProductSelector` ea28ec2
+* add `productSlice` 9932dbc
+* add `RTK` and `axios` f80bd3f
+* add `submitCart` d90136f
+* add api gateway c414c40
+* add item to cart e0712ef
+* add more information on select location 35dbfab
+* add responsive dd81adf
+* add validation on `addItem` 0c924e7
+* disabled qty input on cart 5374cb5
+* render gray marker for selected location 17278cb
+
+
+### Bug Fixes
+
+* `Map` warning on invalid type of var 3716b3c
+* `productSlice` mutate and selector 1e3fa81
+* invalid import ref from `LocationInput` 91a079c
+
 ### [0.1.3](///compare/v0.1.2...v0.1.3) (2021-03-18)